requirements.txt
openai-whisper
tensorflow
transformers
